requirements.txt
Flask==3.1.0
anthropic>=0.42.0
opencv-python==4.11.0.86
Pillow==11.1.0
python-dotenv==1.0.1
gunicorn==23.0.0
flask-limiter==3.9.0
